Add Core Update API
